TileDoc Reference bugfix, Layout fixes, EditProposal Messenger integration
