+++ packages/app/src/components/bars/sidebar/proposal/ProposalStartFundingControl.tsx
@@ -14,7 +14,7 @@ import { useState } from 'react'
 
 const ProposalStartFundingControl = () => {
     const { currentUser } = useCurrentUser()
-    const { proposalStack, proposalStreamDoc, updateProposal } = useProposal()
+    const { proposalStack, proposalStreamID, updateProposal } = useProposal()
     const [isInTransaction, setIsInTransaction] = useState(false)
     const [errorMessage, setErrorMessage] = useState<ErrorMessageType>()
 
@@ -22,12 +22,11 @@ const ProposalStartFundingControl = () => {
         setIsInTransaction(true)
         try {
             if (!currentUser) throw GENERAL_ERROR['NO_WALLET_CONNECTION']
-            if (!proposalStack || !proposalStreamDoc)
-                throw GENERAL_ERROR['CERAMIC_LOAD_ERROR']
+            if (!proposalStack) throw GENERAL_ERROR['CERAMIC_LOAD_ERROR']
             const ceramicStagehand = new CeramicStagehand(currentUser.selfID)
             await ceramicStagehand.deployProposal(
                 currentUser,
-                proposalStreamDoc,
+                proposalStreamID,
                 proposalStack
             )
             await updateProposal()
