+++ packages/app/src/ui/proposals/EditProposalUI.tsx
@@ -1,8 +1,9 @@
-import { Box, Tab, Tabs } from 'grommet'
+import { Box, Stack, Tab, Tabs } from 'grommet'
 
 import Custom404Page from 'packages/app/pages/404'
 import { LOADING_MESSAGE } from '@cambrian/app/constants/LoadingMessages'
 import LoadingScreen from '@cambrian/app/components/info/LoadingScreen'
+import Messenger from '@cambrian/app/components/messenger/Messenger'
 import PageLayout from '@cambrian/app/components/layout/PageLayout'
 import ProposalDescriptionForm from './forms/ProposalDescriptionForm'
 import ProposalFlexInputsForm from './forms/ProposalFlexInputsForm'
@@ -10,6 +11,7 @@ import ProposalHeader from '@cambrian/app/components/layout/header/ProposalHeade
 import ProposalPricingForm from './forms/ProposalPricingForm'
 import { ProposalStatus } from '@cambrian/app/models/ProposalStatus'
 import ProposalSubmitControl from '@cambrian/app/components/bars/sidebar/proposal/ProposalSubmitControl'
+import { useCurrentUser } from '@cambrian/app/hooks/useCurrentUser'
 import useEditProposal from '@cambrian/app/hooks/useEditProposal'
 import { useState } from 'react'
 
@@ -21,7 +23,10 @@ const EditProposalUI = () => {
         setProposalInput,
         onSaveProposal,
         onResetProposalInput,
+        proposalStreamDoc,
     } = useEditProposal()
+
+    const { currentUser } = useCurrentUser()
     const [activeIndex, setActiveIndex] = useState(0)
 
     const isEditable =
@@ -34,48 +39,24 @@ const EditProposalUI = () => {
             {!proposalStack ? (
                 <LoadingScreen context={LOADING_MESSAGE['PROPOSAL']} />
             ) : proposalInput && isEditable ? (
-                <PageLayout contextTitle={'Edit Proposal'} kind="narrow">
-                    <Box pad="large" gap="medium">
-                        <ProposalHeader
-                            proposalStatus={proposalStatus}
-                            proposalStack={proposalStack}
-                        />
-                        <ProposalSubmitControl />
-                        <Tabs
-                            justify="start"
-                            activeIndex={activeIndex}
-                            onActive={(nextIndex: number) =>
-                                setActiveIndex(nextIndex)
-                            }
-                        >
-                            <Tab title="Description">
-                                <Box pad="small">
-                                    <ProposalDescriptionForm
-                                        proposalInput={proposalInput}
-                                        setProposalInput={setProposalInput}
-                                        onSubmit={onSaveProposal}
-                                        onCancel={onResetProposalInput}
-                                    />
-                                </Box>
-                            </Tab>
-                            <Tab title="Pricing">
-                                <Box pad="small">
-                                    <ProposalPricingForm
-                                        template={proposalStack.template}
-                                        proposalInput={proposalInput}
-                                        setProposalInput={setProposalInput}
-                                        onSubmit={onSaveProposal}
-                                        onCancel={onResetProposalInput}
-                                    />
-                                </Box>
-                            </Tab>
-                            {proposalInput.flexInputs.length > 0 && (
-                                <Tab title="Solver Config">
-                                    <Box pad="small">
-                                        <ProposalFlexInputsForm
-                                            composition={
-                                                proposalStack.composition
-                                            }
+                <Stack anchor="bottom-right">
+                    <PageLayout contextTitle={'Edit Proposal'} kind="narrow">
+                        <Box pad="large" gap="medium">
+                            <ProposalHeader
+                                proposalStatus={proposalStatus}
+                                proposalStack={proposalStack}
+                            />
+                            <ProposalSubmitControl />
+                            <Tabs
+                                justify="start"
+                                activeIndex={activeIndex}
+                                onActive={(nextIndex: number) =>
+                                    setActiveIndex(nextIndex)
+                                }
+                            >
+                                <Tab title="Description">
+                                    <Box pad={{ top: 'medium' }}>
+                                        <ProposalDescriptionForm
                                             proposalInput={proposalInput}
                                             setProposalInput={setProposalInput}
                                             onSubmit={onSaveProposal}
@@ -83,10 +64,50 @@ const EditProposalUI = () => {
                                         />
                                     </Box>
                                 </Tab>
-                            )}
-                        </Tabs>
-                    </Box>
-                </PageLayout>
+                                <Tab title="Pricing">
+                                    <Box pad={{ top: 'medium' }}>
+                                        <ProposalPricingForm
+                                            template={proposalStack.template}
+                                            proposalInput={proposalInput}
+                                            setProposalInput={setProposalInput}
+                                            onSubmit={onSaveProposal}
+                                            onCancel={onResetProposalInput}
+                                        />
+                                    </Box>
+                                </Tab>
+                                {proposalInput.flexInputs.length > 0 && (
+                                    <Tab title="Solver Config">
+                                        <Box pad={{ top: 'medium' }}>
+                                            <ProposalFlexInputsForm
+                                                composition={
+                                                    proposalStack.composition
+                                                }
+                                                proposalInput={proposalInput}
+                                                setProposalInput={
+                                                    setProposalInput
+                                                }
+                                                onSubmit={onSaveProposal}
+                                                onCancel={onResetProposalInput}
+                                            />
+                                        </Box>
+                                    </Tab>
+                                )}
+                            </Tabs>
+                        </Box>
+                    </PageLayout>
+                    {proposalStatus !== ProposalStatus.Draft &&
+                        currentUser &&
+                        proposalStreamDoc && (
+                            <Messenger
+                                currentUser={currentUser}
+                                chatID={proposalStreamDoc.id.toString()}
+                                participantDIDs={[
+                                    proposalStack.template.author,
+                                ]}
+                                chatType={'Draft'}
+                            />
+                        )}
+                </Stack>
             ) : (
                 <Custom404Page />
             )}
